Promote TestProject to lib/
So it can be reused in other clients.
